refactor: remove `cmd/igox` in favor of `ispx`
The `cmd/igox` package was a legacy interpreter runtime that has been
superseded by the new `ispx` module. This consolidates the build
infrastructure by moving the build script to `ispx` and creating a new
PC entry point at `ispx/cmd/ispxpc`.

Changes:
- Move `cmd/igox/build.sh` to `ispx/build.sh`
- Add `ispx/cmd/ispxpc/main.go` as PC shared library entry point
- Update `cmd/gox/install.sh` to use `ispx/build.sh`
- Remove unused `SetupWeb` function from `cmd/gox/pkg/command/env.go`
- Delete `cmd/igox` directory entirely

The `ispx` module provides a cleaner, more modern implementation with
better separation of concerns.
